CI: Fix build failure when using ccache+GCC-9
Configure ccache before building to NOT strip comments.

The GCC-compatible way to indicate that a case statement inside a switch
is supposed to fall thru is to add a line at the bottom of the case
block "// fall through" - obviously this is a comment, and will cause a
warning if ccache strips it out.